% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.lrnarg.R
\name{add.lrnarg}
\alias{add.lrnarg}
\title{This function will add a new list of argument to the current object given as parameter.}
\usage{
\method{add}{lrnarg}(object, layer, param, ...)
}
\arguments{
\item{object}{The current lrnarg object.}

\item{layer}{New layer name to be added.}

\item{param}{New parameter list.}

\item{...}{More arguments to be passed to further function.}
}
\value{
Object from class \code{lrnarg}.
}
\description{
This function will add a new list of argument to the current object given as parameter.
}
\examples{
test_lrnarg <- lrnarg(object = list(probability = TRUE),
                          layer = "methylation")
test_lrnarg <- add(object = test_lrnarg,
                   layer = "genexpr",
                   list(probability = TRUE))
}
