% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.methods.R
\name{add}
\alias{add}
\title{Generic function to add element to the object passed as argument.}
\usage{
add(object, ...)
}
\arguments{
\item{object}{The given object from class \code{entity}, \code{lnrarg} or \code{varselectarg}.}

\item{...}{More arguments to be passed to further function.}
}
\description{
Generic function to add element to the object passed as argument.
}
