% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestSpecificLearner.R
\name{bestSpecificLearner}
\alias{bestSpecificLearner}
\title{The best layer-specific model is used as meta model.}
\usage{
bestSpecificLearner(x, y, perf = NULL)
}
\arguments{
\item{x}{\code{data.frame(1)} \cr
\code{data.frame} of predictors.}

\item{y}{\code{vector(1)} \cr
Target observations. Either binary or two level factor variable.}

\item{perf}{\verb{function(1)} \cr
Function to compute layer-specific performance of learners. If NULL, the Brier Score is used by default as performance measure.
Otherwise, the performance function must accept two parameters: \code{observed} (observed values) and \code{predicted} (predicted values).}
}
\value{
A model object of class \code{weightedMeanLeaner}.
}
\description{
The meta learner is the best layer-specific laerner.
}
\examples{
set.seed(20240624L)
x = data.frame(x1 = runif(n = 50L, min = 0, max = 1))
y = sample(x = 0L:1L, size = 50L, replace = TRUE)
my_best_model = bestSpecificLearner(x = x, y = y)

}
