% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.calib.weights.R
\name{build_calib_weights}
\alias{build_calib_weights}
\title{Build Weights with calibration for COBRA Predictions}
\usage{
build_calib_weights(train, test, n_train, n_test, nlearners, eps, alpha)
}
\arguments{
\item{train}{Training data. Rows represent observations,
and columns contain predictions from individual learners for these observations.
In cases where a prediction is unavailable for a specific observation, \code{NA}
is used.}

\item{test}{Test data. Rows represent observations,
and columns contain predictions from individual learners for these observations.
In cases where a prediction is unavailable for a specific observation, \code{NA}
is used.}

\item{n_train}{An \code{integer} specifying the number of training observations.}

\item{n_test}{An \code{integer} specifying the number of test observations.}

\item{nlearners}{An \code{integer} representing the number of learners.}

\item{eps}{A \code{numeric} value representing the threshold for proximity between two predictions.}

\item{alpha}{A value that determines the optimal number of learners in the neighborhood.}
}
\description{
The \code{build_calib_weights} function is used to calculate weights
for predictions when (\code{calib = TRUE}).
}
