% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.dif.R
\name{build_dif}
\alias{build_dif}
\title{Build Difference}
\usage{
build_dif(x)
}
\arguments{
\item{x}{Predictions vector}
}
\description{
The \code{build_dif} function computes the difference between the maximum and minimum predictions in a dataset.
}
