% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.risk.R
\name{build_risk}
\alias{build_risk}
\title{Build Risk}
\usage{
build_risk(pred, target)
}
\arguments{
\item{pred}{A \code{vector} of predictions.}

\item{target}{A \code{vector} of target values.}
}
\description{
Build Risk
}
