% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobra.R
\name{cobra}
\alias{cobra}
\title{Cobra Meta Learner}
\usage{
cobra(x, y, calib = FALSE, eps = NULL)
}
\arguments{
\item{x}{A \code{data.frame} containing the training dataset. Rows represent observations,
and columns contain predictions from individual learners for these observations.
In cases where a prediction is unavailable for a specific observation, \code{NA} is used.}

\item{y}{A \code{vector} containing the training targets. This can be a binary or two-level factor variable.}

\item{calib}{A \code{logical} value indicating whether to perform calibration to find the optimal
predictions. If \code{TRUE}, the COBRA algorithm searches for the optimal \code{alpha}
and \code{eps} as described in the paper. If \code{FALSE}, COBRA is used without calibration.
It is recommended to set \code{calib = FALSE} in the presence of missing values.
The default value is \code{FALSE}.}

\item{eps}{A \code{numeric} value representing the threshold for proximity between two predictions.
The default value is \code{0.1}. It is only used if \code{calib}
is set to \code{FALSE}, allowing users to define the proximity of two predictions.}
}
\value{
A object of class \code{cobra}.
}
\description{
The \code{cobra} function implements the COBRA (COmBined Regression Alternative)
aggregation method, a nonlinear approach for combining predictions
from multiple single learners. This method aims to find optimal predictions
by averaging the target values of the nearest candidates in the training
dataset's predictions. Only the training points that are within the s
pecified proximity to the test point are used to determine the prediction.
If no suitable training points are found, the function returns \code{NA} as the
prediction.
}
\details{
When calibration is performed (\code{calib = TRUE}), the COBRA algorithm
automatically searches for the optimal \code{alpha} (the number of learners whose
predictions are in the neighborhood) and \code{eps} (the threshold for the proximity
of two predictions). In the absence of calibration (\code{calib = FALSE}), \code{eps} defaults
to \code{0.1}
unless otherwise specified by the user, and the number of nearest training points
is determined for each test point.
}
\examples{
# Create a simple training dataset
set.seed(123)
x_train <- data.frame(a = rnorm(10L), b = rnorm(10L))
y_train <- sample(x = 0L:1L, size = 10L, replace = TRUE)

# Train the model with calibration
cobra_model <- cobra(x = x_train, y = y_train, calib = TRUE)

# Make predictions on new data
set.seed(156)
x_new <- data.frame(a = rnorm(5L), b = rnorm(5L))
prediction <- predict(object = cobra_model, data = x_new)
}
\references{
Biau, G., Fischer, A., Guedj, B., & Malley, J. D. (2014). COBRA: A combined regression strategy.
}
