% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.dif.R
\name{create_dif}
\alias{create_dif}
\title{Create Difference}
\usage{
create_dif(x)
}
\arguments{
\item{x}{Predictions vector}
}
\description{
The \code{create_dif} function computes the difference between the maximum and minimum predictions in a dataset.
}
