% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.loss.R
\name{create_loss}
\alias{create_loss}
\title{Create Loss}
\usage{
create_loss(pred, target)
}
\arguments{
\item{pred}{A \code{vector} of predictions.}

\item{target}{A \code{vector} of target values.}
}
\description{
Create Loss
}
