% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDif.R
\name{createDif}
\alias{createDif}
\title{Create Difference}
\usage{
createDif(x)
}
\arguments{
\item{x}{Predictions vector}
}
\description{
The \code{createDif} function computes the difference between the maximum and minimum predictions in a dataset.
}
