% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLoss.R
\name{createLoss}
\alias{createLoss}
\title{Create Loss}
\usage{
createLoss(pred, target)
}
\arguments{
\item{pred}{A \code{vector} of predictions.}

\item{target}{A \code{vector} of target values.}
}
\description{
Create Loss
}
