% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTesting.R
\name{createTesting}
\alias{createTesting}
\title{createTesting}
\usage{
createTesting(id, ind_col)
}
\arguments{
\item{id}{(\code{character(1)}) \cr
Testing id.}

\item{ind_col}{(\code{character(1)}) \cr
Name of column of individuals IDs in testing data.frame.}
}
\value{
A \link{Testing} object.
}
\description{
Creates a \link{Testing} object.
}
