% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTesting.R
\name{createTestLayer}
\alias{createTestLayer}
\title{Creates and store a \link{TestLayer} on the \link{Testing} object passed as argument.}
\usage{
createTestLayer(testing, test_layer_id, test_data)
}
\arguments{
\item{testing}{(\code{Testing(1)}) \cr
Testing object where the created layer will be stored.}

\item{test_layer_id}{(\code{character(1)}) \cr
ID of the testing layer to be created.}

\item{test_data}{(\code{data.frame(1)}) \cr
Data modality to be stored in \link{TestData}.}
}
\value{
The updated \link{Testing} object (with the new layer) is returned.
}
\description{
Creates and store a \link{TestLayer} on the \link{Testing} object passed as argument.
}
