% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\docType{data}
\name{disease}
\alias{disease}
\title{Simulated disease for 200 participants approximately balanced into negative and positive groups.}
\format{
A vector of patient status.
}
\usage{
data(disease)
}
\description{
A dataset containing simulated entities data using the R package \code{interSIM}.
}
\keyword{datasets}
