% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\docType{data}
\name{entities}
\alias{entities}
\title{Simulated multiomics data for 200 participants the same effect size (20) on each layer.}
\format{
A list with three data entities: methylation, gene expressions and protein expressions.
}
\usage{
data(entities)
}
\description{
A dataset containing simulated entities data using the R package \code{interSIM}.
}
\details{
\itemize{
  \item \code{methylation}: A matrix containing the simulated methylation dataset.
  \item \code{genexpr}    : A matrix containing the gene expression dataset.
  \item \code{proteinexpr}: A matrix containing the protein expression dataset.
}
}
\keyword{datasets}
