% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.entitty.R
\name{entity}
\alias{entity}
\title{Constructor of entity object}
\usage{
entity(object, layer)
}
\arguments{
\item{object}{Data entity as \code{matrix} or \code{array}.}

\item{layer}{Layer label.}
}
\value{
Returns an object of class \code{entity}.
}
\description{
Constructor of entity object
}
\examples{
data(entities)
test_entities <- entity(object = entities$methylation,
                          layer = "methylation")
}
