% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{extractData}
\alias{extractData}
\title{extractData}
\usage{
extractData(training)
}
\arguments{
\item{training}{(\code{Training(1)}) \cr
The \link{Training} object of interest.}
}
\value{
A list of data is returned.
}
\description{
Extracts data stored on each layers; base and meta data are extracted.
}
