% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{extractModel}
\alias{extractModel}
\title{extractModel}
\usage{
extractModel(training)
}
\arguments{
\item{training}{(\code{Training(1)}) \cr
The \link{Training} object of interest.}
}
\value{
A list of models is returned.
}
\description{
Extracts models stored on each layers; base and meta models are extracted.
}
