% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layer.R
\name{Layer}
\alias{Layer}
\title{Layer Class}
\description{
This class implements a layer. A \link{Layer} object can only exist as a component of a \link{Study} object.

A layer is structured as followed:
\itemize{
\item \link{Lrner}: It is set by the user to be trained on the training data.
\item \link{TrainData}: It is set by the user to be used to train the learner.
\item \link{Model}: The result of training the learner on the training data, and therefore, not set by the user.
\item \link{NewData}: It is set by the user implements new data to be predicted.
}

A layer can train its learner on its training data and store the resulting model. See the public function \code{Layer$train()} below.

A layer can make predictions for a new layer passed as argument to its predict function. See the public function \code{Layer$predict()} below.
}
\seealso{
\link{Study}, \link{Lrner}, \link{TrainData}, \link{NewData} and \link{Model}
}
\section{Super class}{
\code{\link[fuseMLR:HashTable]{fuseMLR::HashTable}} -> \code{Layer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Layer-new}{\code{Layer$new()}}
\item \href{#method-Layer-print}{\code{Layer$print()}}
\item \href{#method-Layer-getStudy}{\code{Layer$getStudy()}}
\item \href{#method-Layer-train}{\code{Layer$train()}}
\item \href{#method-Layer-predict}{\code{Layer$predict()}}
\item \href{#method-Layer-getTrainData}{\code{Layer$getTrainData()}}
\item \href{#method-Layer-getTargetValues}{\code{Layer$getTargetValues()}}
\item \href{#method-Layer-getIndIDs}{\code{Layer$getIndIDs()}}
\item \href{#method-Layer-getNewData}{\code{Layer$getNewData()}}
\item \href{#method-Layer-getLrner}{\code{Layer$getLrner()}}
\item \href{#method-Layer-getModel}{\code{Layer$getModel()}}
\item \href{#method-Layer-getPredictions}{\code{Layer$getPredictions()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="add2HashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-add2HashTable'><code>fuseMLR::HashTable$add2HashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getFromHashTable'><code>fuseMLR::HashTable$getFromHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getHashTable'><code>fuseMLR::HashTable$getHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getId"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getId'><code>fuseMLR::HashTable$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getKeyClass"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getKeyClass'><code>fuseMLR::HashTable$getKeyClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="removeFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-removeFromHashTable'><code>fuseMLR::HashTable$removeFromHashTable()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$new(id, study)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character()})\cr
See class Param}

\item{\code{study}}{(\code{Study()})\cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-print"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-getStudy"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-getStudy}{}}}
\subsection{Method \code{getStudy()}}{
Getter of the current study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$getStudy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current study is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-train"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-train}{}}}
\subsection{Method \code{train()}}{
Trains the current layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$train(ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ind_subset}}{\link{vector} \cr
ID subset of individuals to be used for training.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The current layer is returned with the resulting model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-predict}{}}}
\subsection{Method \code{predict()}}{
Predicts values for the new layer taking as argument.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$predict(new_layer, ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_layer}}{\code{\link[=Layer]{Layer()}} \cr}

\item{\code{ind_subset}}{\code{\link[=vector]{vector()}} \cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \link{Study} object with the predicted values is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-getTrainData"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-getTrainData}{}}}
\subsection{Method \code{getTrainData()}}{
Getter of the training dataset stored on the current layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$getTrainData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored \link{TrainData} object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-getTargetValues"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-getTargetValues}{}}}
\subsection{Method \code{getTargetValues()}}{
Getter of target values from the current layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$getTargetValues()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing individuals IDs and corresponding target values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-getIndIDs"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-getIndIDs}{}}}
\subsection{Method \code{getIndIDs()}}{
Getter of IDS from the current layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$getIndIDs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing individuals IDs values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-getNewData"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-getNewData}{}}}
\subsection{Method \code{getNewData()}}{
Getter of the new data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$getNewData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored \link{NewData} object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-getLrner"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-getLrner}{}}}
\subsection{Method \code{getLrner()}}{
Getter of the learner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$getLrner()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored \link{Lrner} object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-getModel"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-getModel}{}}}
\subsection{Method \code{getModel()}}{
Getter of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$getModel()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored \link{Model} object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layer-getPredictions"></a>}}
\if{latex}{\out{\hypertarget{method-Layer-getPredictions}{}}}
\subsection{Method \code{getPredictions()}}{
Getter of predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$getPredictions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored predictions are returned.
}
}
}
