% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.lrnarg.R
\name{lrnarg}
\alias{lrnarg}
\title{Constructor of argument object to be used by the learner at each layer.}
\usage{
lrnarg(object, layer)
}
\arguments{
\item{object}{Arguments as \code{list}.}

\item{layer}{Layer label.}
}
\value{
Returns an object of class \code{lrnarg}.
}
\description{
Constructor of argument object to be used by the learner at each layer.
}
\examples{
test_lrnarg <- lrnarg(object = list(probability = TRUE),
                          layer = "methylation")
}
