% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrner.R
\name{Lrner}
\alias{Lrner}
\title{Lrner Class}
\description{
This class implements a learner. A \link{Lrner} object can only exist as a component of a
\link{TrainLayer} or a \link{TrainMetaLayer} object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lrner-new}{\code{Lrner$new()}}
\item \href{#method-Lrner-print}{\code{Lrner$print()}}
\item \href{#method-Lrner-train}{\code{Lrner$train()}}
\item \href{#method-Lrner-getTrainLayer}{\code{Lrner$getTrainLayer()}}
\item \href{#method-Lrner-getId}{\code{Lrner$getId()}}
\item \href{#method-Lrner-getPackage}{\code{Lrner$getPackage()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrner-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lrner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a default parameters list.

Learner ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrner$new(id, package = NULL, lrn_fct, param, train_layer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character()}) \cr
Package that implements the learn function. If NULL, the
learn function is called from the current environment.}

\item{\code{package}}{(\code{character()}) \cr
Learn function name.}

\item{\code{lrn_fct}}{(\code{character()}) \cr
Learn parameters.}

\item{\code{param}}{(\code{Param()}) \cr
Layer on which the learner is stored.}

\item{\code{train_layer}}{(\code{TrainLayer()}) \cr
The training layer where to store the learner.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrner-print"></a>}}
\if{latex}{\out{\hypertarget{method-Lrner-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrner$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrner-train"></a>}}
\if{latex}{\out{\hypertarget{method-Lrner-train}{}}}
\subsection{Method \code{train()}}{
Tains the current learner (from class \link{Lrner}) on the current training data (from class \link{TrainData}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrner$train(ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ind_subset}}{\cr
Individual ID subset on which the training will be performed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The resulting model, from class \link{Model}, is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrner-getTrainLayer"></a>}}
\if{latex}{\out{\hypertarget{method-Lrner-getTrainLayer}{}}}
\subsection{Method \code{getTrainLayer()}}{
The current layer is returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrner$getTrainLayer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link{TrainLayer} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrner-getId"></a>}}
\if{latex}{\out{\hypertarget{method-Lrner-getId}{}}}
\subsection{Method \code{getId()}}{
Getter of the current learner ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrner$getId()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current learner ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrner-getPackage"></a>}}
\if{latex}{\out{\hypertarget{method-Lrner-getPackage}{}}}
\subsection{Method \code{getPackage()}}{
Getter of the learner package implementing the learn function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrner$getPackage()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The name of the package implementing the learn function.
}
}
}
