% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaLayer.R
\name{MetaLayer}
\alias{MetaLayer}
\title{MetaLayer Class}
\description{
This class implement a meta meta layer. A \link{MetaLayer} can only exist as unique element of a \link{Study} object.

A layer is structured as followed:
\itemize{
\item \link{Lrner}: It is set by the user to be trained on the meta training data.
\item \link{TrainData}: It are meta data, automatically created by the internal cross validation.
\item \link{Model}: The meta model, result of training the learner on the training data, and therefore, not to be set by the user.
\item \link{NewData}: The meta new data to be predicted, consisting in predictions obtained from each layer.
}

A meta layer can train its meta learner on the meta training data and store the resulting meta model.
The meta layer can predict values given a new meta layer.
}
\section{Super class}{
\code{\link[fuseMLR:HashTable]{fuseMLR::HashTable}} -> \code{MetaLayer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MetaLayer-new}{\code{MetaLayer$new()}}
\item \href{#method-MetaLayer-print}{\code{MetaLayer$print()}}
\item \href{#method-MetaLayer-getStudy}{\code{MetaLayer$getStudy()}}
\item \href{#method-MetaLayer-train}{\code{MetaLayer$train()}}
\item \href{#method-MetaLayer-predict}{\code{MetaLayer$predict()}}
\item \href{#method-MetaLayer-getTrainData}{\code{MetaLayer$getTrainData()}}
\item \href{#method-MetaLayer-getNewData}{\code{MetaLayer$getNewData()}}
\item \href{#method-MetaLayer-getLrner}{\code{MetaLayer$getLrner()}}
\item \href{#method-MetaLayer-getModel}{\code{MetaLayer$getModel()}}
\item \href{#method-MetaLayer-openAccess}{\code{MetaLayer$openAccess()}}
\item \href{#method-MetaLayer-closeAccess}{\code{MetaLayer$closeAccess()}}
\item \href{#method-MetaLayer-getAccess}{\code{MetaLayer$getAccess()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="add2HashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-add2HashTable'><code>fuseMLR::HashTable$add2HashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getFromHashTable'><code>fuseMLR::HashTable$getFromHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getHashTable'><code>fuseMLR::HashTable$getHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getId"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getId'><code>fuseMLR::HashTable$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getKeyClass"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getKeyClass'><code>fuseMLR::HashTable$getKeyClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="removeFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-removeFromHashTable'><code>fuseMLR::HashTable$removeFromHashTable()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-new"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$new(id, study)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character()})\cr
See class Param}

\item{\code{study}}{(\code{Study()})\cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-print"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-getStudy"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-getStudy}{}}}
\subsection{Method \code{getStudy()}}{
Getter of the current study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$getStudy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current study is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-train"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-train}{}}}
\subsection{Method \code{train()}}{
Trains the current layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$train(ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ind_subset}}{\link{vector} \cr
ID subset of individuals to be used for training.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The current layer is returned with the resulting model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-predict"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-predict}{}}}
\subsection{Method \code{predict()}}{
Predicts values for the new layer taking as argument.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$predict(new_layer, ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_layer}}{\code{\link[=Layer]{Layer()}} \cr}

\item{\code{ind_subset}}{\code{\link[=vector]{vector()}} \cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new study with the predicted values is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-getTrainData"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-getTrainData}{}}}
\subsection{Method \code{getTrainData()}}{
Getter of the training dataset stored on the current layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$getTrainData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored \link{TrainData} object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-getNewData"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-getNewData}{}}}
\subsection{Method \code{getNewData()}}{
Getter of the new data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$getNewData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored \link{NewData} object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-getLrner"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-getLrner}{}}}
\subsection{Method \code{getLrner()}}{
Getter of the learner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$getLrner()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored \link{Lrner} object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-getModel"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-getModel}{}}}
\subsection{Method \code{getModel()}}{
Getter of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$getModel()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored \link{Model} object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-openAccess"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-openAccess}{}}}
\subsection{Method \code{openAccess()}}{
Open access to the meta layer. A meta learner is only
modifiable if the access is opened.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$openAccess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-closeAccess"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-closeAccess}{}}}
\subsection{Method \code{closeAccess()}}{
Close access to the meta layer to avoid accidental
modification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$closeAccess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MetaLayer-getAccess"></a>}}
\if{latex}{\out{\hypertarget{method-MetaLayer-getAccess}{}}}
\subsection{Method \code{getAccess()}}{
Getter of the current access to the meta layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaLayer$getAccess()}\if{html}{\out{</div>}}
}

}
}
