% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVarSelect.R
\name{multiVarSelect}
\alias{multiVarSelect}
\title{Variable selection is conducted for the provided function on each data entity.}
\usage{
multiVarSelect(data, target, var_selec = "Boruta", var_selec_arg = list())
}
\arguments{
\item{data}{A list containing data entities.}

\item{target}{The target variables.}

\item{var_selec}{The function name to be called for variable selection.}

\item{var_selec_arg}{Arguments to be passed to the variable selection function.}
}
\value{
Object of class \code{multiVarSelect} containing following information on each layer:
\tabular{ll}{
   \code{learner} \tab Name of the learner, e.g. \code{ranger},\cr
   \code{var_selected} \tab Information about variable selection.\cr
 }
}
\description{
Variable selection is conducted for the provided function on each data entity.
}
\examples{
\dontrun{
data(entities)
data(disease)
multiVarSelect(data = entities,
               target = disease,
               var_selec = "Boruta_ext",
               var_selec_arg = list())
## With the original Boruta function
multiVarSelect(data = entities,
target = disease,
var_selec = "Boruta",
var_selec_arg = list(list(mtry = 2),
                     list(mtry = 2),
                     list(mtry = 2)))
}
}
