% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewData.R
\name{NewData}
\alias{NewData}
\title{NewData Class}
\description{
This class implements \link{NewData} object to be predicted.
A \link{NewData} object can only exist as a component of a \link{NewLayer} or a \link{NewMetaLayer} object.
}
\seealso{
\link{TrainData}
}
\section{Super class}{
\code{fuseMLR::Data} -> \code{NewData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NewData-new}{\code{NewData$new()}}
\item \href{#method-NewData-print}{\code{NewData$print()}}
\item \href{#method-NewData-getData}{\code{NewData$getData()}}
\item \href{#method-NewData-getNewLayer}{\code{NewData$getNewLayer()}}
\item \href{#method-NewData-clone}{\code{NewData$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Data" data-id="getCompleteData"><a href='../../fuseMLR/html/Data.html#method-Data-getCompleteData'><code>fuseMLR::Data$getCompleteData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Data" data-id="getDataFrame"><a href='../../fuseMLR/html/Data.html#method-Data-getDataFrame'><code>fuseMLR::Data$getDataFrame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Data" data-id="getId"><a href='../../fuseMLR/html/Data.html#method-Data-getId'><code>fuseMLR::Data$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Data" data-id="getIndCol"><a href='../../fuseMLR/html/Data.html#method-Data-getIndCol'><code>fuseMLR::Data$getIndCol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Data" data-id="getIndSubset"><a href='../../fuseMLR/html/Data.html#method-Data-getIndSubset'><code>fuseMLR::Data$getIndSubset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Data" data-id="getSetDiff"><a href='../../fuseMLR/html/Data.html#method-Data-getSetDiff'><code>fuseMLR::Data$getSetDiff()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Data" data-id="getVarSubset"><a href='../../fuseMLR/html/Data.html#method-Data-getVarSubset'><code>fuseMLR::Data$getVarSubset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Data" data-id="setDataFrame"><a href='../../fuseMLR/html/Data.html#method-Data-setDataFrame'><code>fuseMLR::Data$setDataFrame()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewData-new"></a>}}
\if{latex}{\out{\hypertarget{method-NewData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new object from the current class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewData$new(id, data_frame, new_layer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)}) \cr
Object ID.}

\item{\code{data_frame}}{(\code{data.frame(1)})\cr
\code{data.frame} containing data.}

\item{\code{new_layer}}{(\code{NewLayer(1)}) \cr
Layer where to store the current object.}

\item{\code{ind_col}}{(\code{character(1)})\cr
Column name containing individual IDs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewData-print"></a>}}
\if{latex}{\out{\hypertarget{method-NewData-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewData$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewData-getData"></a>}}
\if{latex}{\out{\hypertarget{method-NewData-getData}{}}}
\subsection{Method \code{getData()}}{
Getter of the current \code{data.frame} wihtout individual
ID variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewData$getData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The \code{data.frame} without individual ID nor target variables is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewData-getNewLayer"></a>}}
\if{latex}{\out{\hypertarget{method-NewData-getNewLayer}{}}}
\subsection{Method \code{getNewLayer()}}{
Getter of the current layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewData$getNewLayer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The layer (from class \link{NewLayer}) on which the current train data are stored
is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NewData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
