% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewLayer.R
\name{NewLayer}
\alias{NewLayer}
\title{NewLayer Class}
\description{
This class implements a layer. A \link{NewLayer} object can only exist as a component of a \link{PredictStudy} object.

A predicted layer can only contain \link{NewData}.
}
\seealso{
\link{TrainStudy}, \link{Lrner}, \link{TrainData}, \link{NewData} and \link{Model}
}
\section{Super class}{
\code{fuseMLR::HashTable} -> \code{NewLayer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NewLayer-new}{\code{NewLayer$new()}}
\item \href{#method-NewLayer-print}{\code{NewLayer$print()}}
\item \href{#method-NewLayer-getNewStudy}{\code{NewLayer$getNewStudy()}}
\item \href{#method-NewLayer-getIndIDs}{\code{NewLayer$getIndIDs()}}
\item \href{#method-NewLayer-getNewData}{\code{NewLayer$getNewData()}}
\item \href{#method-NewLayer-checkNewDataExist}{\code{NewLayer$checkNewDataExist()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="add2HashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-add2HashTable'><code>fuseMLR::HashTable$add2HashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="checkClassExist"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-checkClassExist'><code>fuseMLR::HashTable$checkClassExist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getFromHashTable'><code>fuseMLR::HashTable$getFromHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getHashTable'><code>fuseMLR::HashTable$getHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getId"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getId'><code>fuseMLR::HashTable$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getKeyClass"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getKeyClass'><code>fuseMLR::HashTable$getKeyClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="removeFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-removeFromHashTable'><code>fuseMLR::HashTable$removeFromHashTable()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewLayer-new"></a>}}
\if{latex}{\out{\hypertarget{method-NewLayer-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewLayer$new(id, new_study)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
See class Param}

\item{\code{new_study}}{(\code{NewStudy(1)})\cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewLayer-print"></a>}}
\if{latex}{\out{\hypertarget{method-NewLayer-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewLayer$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewLayer-getNewStudy"></a>}}
\if{latex}{\out{\hypertarget{method-NewLayer-getNewStudy}{}}}
\subsection{Method \code{getNewStudy()}}{
Getter of the current study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewLayer$getNewStudy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current study is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewLayer-getIndIDs"></a>}}
\if{latex}{\out{\hypertarget{method-NewLayer-getIndIDs}{}}}
\subsection{Method \code{getIndIDs()}}{
Getter of IDS from the current layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewLayer$getIndIDs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing individuals IDs values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewLayer-getNewData"></a>}}
\if{latex}{\out{\hypertarget{method-NewLayer-getNewData}{}}}
\subsection{Method \code{getNewData()}}{
Getter of the predicted data stored on the current layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewLayer$getNewData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored \link{NewData} object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewLayer-checkNewDataExist"></a>}}
\if{latex}{\out{\hypertarget{method-NewLayer-checkNewDataExist}{}}}
\subsection{Method \code{checkNewDataExist()}}{
Check whether a new data has been already stored.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewLayer$checkNewDataExist()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean value
}
}
}
