% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewMetaLayer.R
\name{NewMetaLayer}
\alias{NewMetaLayer}
\title{NewMetaLayer Class}
\description{
This class implement a predicted meta layer. A \link{NewMetaLayer} can only exist as unique element of a \link{TrainStudy} object.

A predicted meta layer can only contain a \link{NewData} object.
}
\section{Super class}{
\code{\link[fuseMLR:HashTable]{fuseMLR::HashTable}} -> \code{NewMetaLayer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NewMetaLayer-new}{\code{NewMetaLayer$new()}}
\item \href{#method-NewMetaLayer-print}{\code{NewMetaLayer$print()}}
\item \href{#method-NewMetaLayer-getNewStudy}{\code{NewMetaLayer$getNewStudy()}}
\item \href{#method-NewMetaLayer-getNewData}{\code{NewMetaLayer$getNewData()}}
\item \href{#method-NewMetaLayer-openAccess}{\code{NewMetaLayer$openAccess()}}
\item \href{#method-NewMetaLayer-closeAccess}{\code{NewMetaLayer$closeAccess()}}
\item \href{#method-NewMetaLayer-getAccess}{\code{NewMetaLayer$getAccess()}}
\item \href{#method-NewMetaLayer-setNewData}{\code{NewMetaLayer$setNewData()}}
\item \href{#method-NewMetaLayer-checkNewDataExist}{\code{NewMetaLayer$checkNewDataExist()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="add2HashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-add2HashTable'><code>fuseMLR::HashTable$add2HashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="checkClassExist"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-checkClassExist'><code>fuseMLR::HashTable$checkClassExist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getFromHashTable'><code>fuseMLR::HashTable$getFromHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getHashTable'><code>fuseMLR::HashTable$getHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getId"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getId'><code>fuseMLR::HashTable$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getKeyClass"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getKeyClass'><code>fuseMLR::HashTable$getKeyClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="removeFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-removeFromHashTable'><code>fuseMLR::HashTable$removeFromHashTable()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewMetaLayer-new"></a>}}
\if{latex}{\out{\hypertarget{method-NewMetaLayer-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewMetaLayer$new(id, new_study)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character()})\cr
See class Param}

\item{\code{new_study}}{(\code{NewStudy()})\cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewMetaLayer-print"></a>}}
\if{latex}{\out{\hypertarget{method-NewMetaLayer-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewMetaLayer$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewMetaLayer-getNewStudy"></a>}}
\if{latex}{\out{\hypertarget{method-NewMetaLayer-getNewStudy}{}}}
\subsection{Method \code{getNewStudy()}}{
Getter of the current predicted study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewMetaLayer$getNewStudy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current new study is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewMetaLayer-getNewData"></a>}}
\if{latex}{\out{\hypertarget{method-NewMetaLayer-getNewData}{}}}
\subsection{Method \code{getNewData()}}{
Getter of the training dataset stored on the current layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewMetaLayer$getNewData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The stored \link{NewData} object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewMetaLayer-openAccess"></a>}}
\if{latex}{\out{\hypertarget{method-NewMetaLayer-openAccess}{}}}
\subsection{Method \code{openAccess()}}{
Open access to the meta layer. A meta learner is only
modifiable if the access is opened.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewMetaLayer$openAccess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewMetaLayer-closeAccess"></a>}}
\if{latex}{\out{\hypertarget{method-NewMetaLayer-closeAccess}{}}}
\subsection{Method \code{closeAccess()}}{
Close access to the meta layer to avoid accidental
modification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewMetaLayer$closeAccess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewMetaLayer-getAccess"></a>}}
\if{latex}{\out{\hypertarget{method-NewMetaLayer-getAccess}{}}}
\subsection{Method \code{getAccess()}}{
Getter of the current access to the meta layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewMetaLayer$getAccess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewMetaLayer-setNewData"></a>}}
\if{latex}{\out{\hypertarget{method-NewMetaLayer-setNewData}{}}}
\subsection{Method \code{setNewData()}}{
Create and set an \link{NewData} object to the current
new meta learner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewMetaLayer$setNewData(id, ind_col, data_frame, meta_layer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character()} \cr
ID of the \link{NewData} object to be instanciated.}

\item{\code{ind_col}}{\code{character()} \cr
Name of individual column IDs.}

\item{\code{data_frame}}{\code{data.frame} \cr
\code{data.frame} of layer specific predictions.}

\item{\code{meta_layer}}{\code{NewLayer()} \cr
Layer where to store the \link{NewData} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewMetaLayer-checkNewDataExist"></a>}}
\if{latex}{\out{\hypertarget{method-NewMetaLayer-checkNewDataExist}{}}}
\subsection{Method \code{checkNewDataExist()}}{
Check whether a new data has been already stored.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewMetaLayer$checkNewDataExist()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean value
}
}
}
