% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewStudy.R
\name{NewStudy}
\alias{NewStudy}
\title{NewStudy Class}
\description{
This class is the basic class of the present package. An object from this class
is designed to contain multiple layers, but only one new meta layer.

A study is structured as followed:
\itemize{
\item \link{NewLayer}
\item \link{NewMetaLayer}
}
}
\seealso{
\link{TrainLayer}
}
\section{Super class}{
\code{\link[fuseMLR:HashTable]{fuseMLR::HashTable}} -> \code{NewStudy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NewStudy-new}{\code{NewStudy$new()}}
\item \href{#method-NewStudy-print}{\code{NewStudy$print()}}
\item \href{#method-NewStudy-createMetaNewData}{\code{NewStudy$createMetaNewData()}}
\item \href{#method-NewStudy-getIndIDs}{\code{NewStudy$getIndIDs()}}
\item \href{#method-NewStudy-getNewMetaLayer}{\code{NewStudy$getNewMetaLayer()}}
\item \href{#method-NewStudy-getIndCol}{\code{NewStudy$getIndCol()}}
\item \href{#method-NewStudy-getTarget}{\code{NewStudy$getTarget()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="add2HashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-add2HashTable'><code>fuseMLR::HashTable$add2HashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="checkClassExist"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-checkClassExist'><code>fuseMLR::HashTable$checkClassExist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getFromHashTable'><code>fuseMLR::HashTable$getFromHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getHashTable'><code>fuseMLR::HashTable$getHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getId"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getId'><code>fuseMLR::HashTable$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getKeyClass"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getKeyClass'><code>fuseMLR::HashTable$getKeyClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="removeFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-removeFromHashTable'><code>fuseMLR::HashTable$removeFromHashTable()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewStudy-new"></a>}}
\if{latex}{\out{\hypertarget{method-NewStudy-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewStudy$new(id, ind_col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
See class Param}

\item{\code{ind_col}}{(\code{character(0L)})
Name of column of individuals IDS}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewStudy-print"></a>}}
\if{latex}{\out{\hypertarget{method-NewStudy-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewStudy$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewStudy-createMetaNewData"></a>}}
\if{latex}{\out{\hypertarget{method-NewStudy-createMetaNewData}{}}}
\subsection{Method \code{createMetaNewData()}}{
Creates a new meta dataset based on layer predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewStudy$createMetaNewData(meta_layer_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meta_layer_id}}{(\code{character()}) \cr
ID of the meta layer where the new meta data will be stored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{NewData} is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewStudy-getIndIDs"></a>}}
\if{latex}{\out{\hypertarget{method-NewStudy-getIndIDs}{}}}
\subsection{Method \code{getIndIDs()}}{
Gather individual IDs from all layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewStudy$getIndIDs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing individuals IDs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewStudy-getNewMetaLayer"></a>}}
\if{latex}{\out{\hypertarget{method-NewStudy-getNewMetaLayer}{}}}
\subsection{Method \code{getNewMetaLayer()}}{
Getter of the meta layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewStudy$getNewMetaLayer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Object from class \link{NewMetaLayer}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewStudy-getIndCol"></a>}}
\if{latex}{\out{\hypertarget{method-NewStudy-getIndCol}{}}}
\subsection{Method \code{getIndCol()}}{
Getter of the individual column name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewStudy$getIndCol()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewStudy-getTarget"></a>}}
\if{latex}{\out{\hypertarget{method-NewStudy-getTarget}{}}}
\subsection{Method \code{getTarget()}}{
Getter of the target variable name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewStudy$getTarget()}\if{html}{\out{</div>}}
}

}
}
