% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param.R
\name{Param}
\alias{Param}
\title{Class Param}
\description{
Implements a parameter set. Objects from this class contain non-tunable parameters.
}
\seealso{
\link{ParamLrner}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Param-new}{\code{Param$new()}}
\item \href{#method-Param-print}{\code{Param$print()}}
\item \href{#method-Param-getId}{\code{Param$getId()}}
\item \href{#method-Param-getParamList}{\code{Param$getParamList()}}
\item \href{#method-Param-clone}{\code{Param$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Param-new"></a>}}
\if{latex}{\out{\hypertarget{method-Param-new}{}}}
\subsection{Method \code{new()}}{
Initialize a default parameters list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param$new(id, param_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
The ID of current parameter object.}

\item{\code{param_list}}{(\code{list(1)})\cr
List of parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Param-print"></a>}}
\if{latex}{\out{\hypertarget{method-Param-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Param-getId"></a>}}
\if{latex}{\out{\hypertarget{method-Param-getId}{}}}
\subsection{Method \code{getId()}}{
Getter of parameter ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param$getId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Param-getParamList"></a>}}
\if{latex}{\out{\hypertarget{method-Param-getParamList}{}}}
\subsection{Method \code{getParamList()}}{
Getter of parameter list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param$getParamList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Param-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Param-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
