% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamLearner.R
\name{ParamLearner}
\alias{ParamLearner}
\title{Class ParamLearner.}
\description{
Implement the list of parameters to be passed to the \link{Lrner} object.
Non-tunable parameters and tunable paramters are stored in the object
from this class.
}
\section{Super class}{
\code{\link[fuseMLR:Param]{fuseMLR::Param}} -> \code{ParamLearner}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParamLearner-new}{\code{ParamLearner$new()}}
\item \href{#method-ParamLearner-print}{\code{ParamLearner$print()}}
\item \href{#method-ParamLearner-getParamLrner}{\code{ParamLearner$getParamLrner()}}
\item \href{#method-ParamLearner-getHyperparam}{\code{ParamLearner$getHyperparam()}}
\item \href{#method-ParamLearner-clone}{\code{ParamLearner$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Param" data-id="getId"><a href='../../fuseMLR/html/Param.html#method-Param-getId'><code>fuseMLR::Param$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Param" data-id="getParamList"><a href='../../fuseMLR/html/Param.html#method-Param-getParamList'><code>fuseMLR::Param$getParamList()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamLearner-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParamLearner-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamLearner$new(id, param_list, hyperparam_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
See class Param}

\item{\code{param_list}}{(\code{list(1)})\cr
See class Param}

\item{\code{hyperparam_list}}{(\code{list(1)})\cr
List of hyperparameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamLearner-print"></a>}}
\if{latex}{\out{\hypertarget{method-ParamLearner-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamLearner$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamLearner-getParamLrner"></a>}}
\if{latex}{\out{\hypertarget{method-ParamLearner-getParamLrner}{}}}
\subsection{Method \code{getParamLrner()}}{
Getter of learner parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamLearner$getParamLrner()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamLearner-getHyperparam"></a>}}
\if{latex}{\out{\hypertarget{method-ParamLearner-getHyperparam}{}}}
\subsection{Method \code{getHyperparam()}}{
Getter of hyperparameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamLearner$getHyperparam()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamLearner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParamLearner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamLearner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
