% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamVarSel.R
\name{ParamVarSel}
\alias{ParamVarSel}
\title{Class ParamVarSel.}
\description{
Implement the list of parameters to be passed to the \link{VarSel} object.
}
\section{Super class}{
\code{fuseMLR::Param} -> \code{ParamVarSel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParamVarSel-new}{\code{ParamVarSel$new()}}
\item \href{#method-ParamVarSel-print}{\code{ParamVarSel$print()}}
\item \href{#method-ParamVarSel-getParamVarSel}{\code{ParamVarSel$getParamVarSel()}}
\item \href{#method-ParamVarSel-clone}{\code{ParamVarSel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Param" data-id="getId"><a href='../../fuseMLR/html/Param.html#method-Param-getId'><code>fuseMLR::Param$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="Param" data-id="getParamList"><a href='../../fuseMLR/html/Param.html#method-Param-getParamList'><code>fuseMLR::Param$getParamList()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamVarSel-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParamVarSel-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamVarSel$new(id, param_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
See class Param}

\item{\code{param_list}}{(\code{list(1)})\cr
See class Param}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamVarSel-print"></a>}}
\if{latex}{\out{\hypertarget{method-ParamVarSel-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamVarSel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamVarSel-getParamVarSel"></a>}}
\if{latex}{\out{\hypertarget{method-ParamVarSel-getParamVarSel}{}}}
\subsection{Method \code{getParamVarSel()}}{
Getter of learner parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamVarSel$getParamVarSel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamVarSel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParamVarSel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamVarSel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
