% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bestSpecificLearner.R
\name{predict.bestSpecificLearner}
\alias{predict.bestSpecificLearner}
\title{Weighted mean prediction.}
\usage{
\method{predict}{bestSpecificLearner}(object, data, na.rm = TRUE)
}
\arguments{
\item{object}{\code{weightedMeanLearner(1)} \cr
An object from class \link{weightedMeanLearner}}

\item{data}{\code{data.frame} \cr
\code{data.frame} to be predicted.}

\item{na.rm}{\cr
Removes NAs when TRUE.}
}
\value{
Predicted target values are returned.
}
\description{
Predict function for models from class \code{weightedMeanLearner}.
}
\examples{
set.seed(20240625)
x = data.frame(x1 = runif(n = 50L, min = 0, max = 1))
y <- sample(x = 0:1, size = 50L, replace = TRUE)
my_model <- bestSpecificLearner(x = x, y = y)
x_new <- data.frame(x1 = rnorm(10L))
my_predictions <- predict(object = my_model, data = x_new)

}
