% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.meta.lnr.R
\name{predict.meta.lnr}
\alias{predict.meta.lnr}
\title{Entity learners will be gathered together as one learner.}
\usage{
\method{predict}{meta.lnr}(object, data, ...)
}
\arguments{
\item{object}{Object of class ranger.cved.}

\item{data}{A list containing data entities.}

\item{...}{To be passed to further function.}
}
\value{
A predictor of class 'predict.ranger.cved'.
}
\description{
Entity learners will be gathered together as one learner.
}
