% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.multiLearner.R
\name{predict.multiLearner}
\alias{predict.multiLearner}
\title{Predict on each layer. The pre-defined predict function of the layer learner should return a list containing a 'predictions' elements. The 'predictions' element contains predicted values.}
\usage{
\method{predict}{multiLearner}(object, data, ...)
}
\arguments{
\item{object}{Object of class 'multiLearner'}

\item{data}{New dataset to be used for predictions.}

\item{...}{Further arguments to be passed to the pre-defined predict function.}
}
\value{
Object of class multiLearner.predict.
}
\description{
Predict on each layer. The pre-defined predict function of the layer learner should return a list containing a 'predictions' elements. The 'predictions' element contains predicted values.
}
