% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ranger.cved.R
\name{predict.ranger.cved}
\alias{predict.ranger.cved}
\title{Entity learners will be gathered together as one learner.}
\usage{
\method{predict}{ranger.cved}(object, data, ...)
}
\arguments{
\item{object}{Object of class ranger.cved.}

\item{data}{A list containing data entities.}

\item{...}{To be passed to further function.}
}
\value{
A predictor of class \code{predict.ranger.cved}.
}
\description{
Entity learners will be gathered together as one learner.
}
