% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{predict.Training}
\alias{predict.Training}
\title{predict.Training}
\usage{
\method{predict}{Training}(object, testing, ind_subset = NULL)
}
\arguments{
\item{object}{\code{Training} \cr
A trained Training object to be used to compute predictions.}

\item{testing}{\code{Testing} \cr
A new testing object to be predicted.}

\item{ind_subset}{\code{vector} \cr
Vector of IDs to be predicted.}
}
\value{
The final predicted object. All layers and the meta layer are predicted.
}
\description{
Computes predictions for the \link{Testing} object passed as argument.
}
