% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.weighted.mean.R
\name{predict.weighted.mean}
\alias{predict.weighted.mean}
\title{Layer learners will be gathered together as one learner.}
\usage{
\method{predict}{weighted.mean}(object, data, ...)
}
\arguments{
\item{object}{Object of class weighted.mean}

\item{data}{A list containing data predicted values.}

\item{...}{To be passed to further function.}
}
\value{
A predictor of class \code{predict.weighted.mean}.
}
\description{
Layer learners will be gathered together as one learner.
}
