% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictStudy.R
\name{PredictStudy}
\alias{PredictStudy}
\title{PredictStudy Class}
\description{
This class is the basic class of the present package. An object from this class
is designed to contain multiple layers, but only one meta layer.

A study is structured as followed:
\itemize{
\item \link{PredictLayer}: Can be clinical, gene expression, etc.
\itemize{
\item \link{Lrner}: Specific to each layer, it must be set up by the user.
\item \link{TrainData}: Specific to each layer, it must be set up by the user.
\item \link{Model}: Specific to each layer, it is set up by training the learner on the training data.
}
\item \link{PredictMetaLayer}: Basically a \link{PredictLayer}, but with some specific properties.
\itemize{
\item \link{Lrner}: This is the meta learner, it must be set up by the user.
\item \link{TrainData}: Specific to each layer, it is set up internally after cross-validation.
\item \link{Model}: Specific to each layer, it is set up by training the learner on the training data.
}
}

Use the function \code{train} to train a study and \code{predict} to predict
a new study.
}
\seealso{
\link{TrainLayer}
}
\section{Super class}{
\code{\link[fuseMLR:HashTable]{fuseMLR::HashTable}} -> \code{PredictStudy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictStudy-new}{\code{PredictStudy$new()}}
\item \href{#method-PredictStudy-print}{\code{PredictStudy$print()}}
\item \href{#method-PredictStudy-createMetaNewData}{\code{PredictStudy$createMetaNewData()}}
\item \href{#method-PredictStudy-getIndIDs}{\code{PredictStudy$getIndIDs()}}
\item \href{#method-PredictStudy-getPredictMetaLayer}{\code{PredictStudy$getPredictMetaLayer()}}
\item \href{#method-PredictStudy-getIndCol}{\code{PredictStudy$getIndCol()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="add2HashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-add2HashTable'><code>fuseMLR::HashTable$add2HashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="checkClassExist"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-checkClassExist'><code>fuseMLR::HashTable$checkClassExist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getFromHashTable'><code>fuseMLR::HashTable$getFromHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getHashTable'><code>fuseMLR::HashTable$getHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getId"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getId'><code>fuseMLR::HashTable$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getKeyClass"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getKeyClass'><code>fuseMLR::HashTable$getKeyClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="removeFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-removeFromHashTable'><code>fuseMLR::HashTable$removeFromHashTable()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictStudy-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictStudy-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictStudy$new(id, ind_col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
See class Param}

\item{\code{ind_col}}{(\code{character(0L)})
Name of column of individuals IDS}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictStudy-print"></a>}}
\if{latex}{\out{\hypertarget{method-PredictStudy-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictStudy$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictStudy-createMetaNewData"></a>}}
\if{latex}{\out{\hypertarget{method-PredictStudy-createMetaNewData}{}}}
\subsection{Method \code{createMetaNewData()}}{
Creates a new meta dataset based on layer predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictStudy$createMetaNewData(meta_layer_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meta_layer_id}}{(\code{character()}) \cr
ID of the meta layer where the new meta data will be stored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{NewData} is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictStudy-getIndIDs"></a>}}
\if{latex}{\out{\hypertarget{method-PredictStudy-getIndIDs}{}}}
\subsection{Method \code{getIndIDs()}}{
Gather individual IDs from all layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictStudy$getIndIDs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing individuals IDs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictStudy-getPredictMetaLayer"></a>}}
\if{latex}{\out{\hypertarget{method-PredictStudy-getPredictMetaLayer}{}}}
\subsection{Method \code{getPredictMetaLayer()}}{
Getter of the meta layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictStudy$getPredictMetaLayer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Object from class \link{PredictMetaLayer}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictStudy-getIndCol"></a>}}
\if{latex}{\out{\hypertarget{method-PredictStudy-getIndCol}{}}}
\subsection{Method \code{getIndCol()}}{
Getter of the individual column name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictStudy$getIndCol()}\if{html}{\out{</div>}}
}

}
}
