% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_boruta.R
\name{print.Boruta_ext}
\alias{print.Boruta_ext}
\title{Print object of class 'Boruta.ext'}
\usage{
\method{print}{Boruta_ext}(x, ...)
}
\arguments{
\item{x}{Object of class Boruta.ext.}

\item{...}{To be used by the further print function.}
}
\description{
Print object of class 'Boruta.ext'
}
