% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.entitty.R
\name{print.entity}
\alias{print.entity}
\title{Print object of class entity.}
\usage{
\method{print}{entity}(x, ...)
}
\arguments{
\item{x}{object to be printed.}

\item{...}{More arguments to be passed to further function.}
}
\description{
Print object of class entity.
}
\examples{
data(entities)
test_entities <- entity(object = entities$methylation,
                          layer = "methylation")
test_entities <- add(object = test_entities,
                   layer = "genexpr",
                   data = entities$genexpr)
print(test_entities)
}
