% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.lrnarg.R
\name{print.lrnarg}
\alias{print.lrnarg}
\title{Print object of class lrnarg.}
\usage{
\method{print}{lrnarg}(x, ...)
}
\arguments{
\item{x}{object to be printed.}

\item{...}{More arguments to be passed to further function.}
}
\description{
Print object of class lrnarg.
}
\examples{
test_lrnarg <- lrnarg(object = list(probability = TRUE),
                          layer = "methylation")
test_lrnarg <- add(object = test_lrnarg,
                   layer = "genexpr",
                   list(probability = TRUE))
print(test_lrnarg)
}
