% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.meta.lnr.R
\name{print.meta.lnr}
\alias{print.meta.lnr}
\title{Print object of class \code{meta.lnr}.}
\usage{
\method{print}{meta.lnr}(x, ...)
}
\arguments{
\item{x}{Object of class \code{meta.lnr}.}

\item{...}{[\code{any}]\cr}
}
\description{
Print object of class \code{meta.lnr}.
}
