% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multiLearner.R
\name{print.multiLearner.predict}
\alias{print.multiLearner.predict}
\title{Print object of class 'multiLearner.predict'}
\usage{
\method{print}{multiLearner.predict}(x, ...)
}
\arguments{
\item{x}{Object of class 'multiLearner.predict'}

\item{...}{[\code{any}]\cr}
}
\description{
Print object of class 'multiLearner.predict'
}
