% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multiLearner.R
\name{print.multiLearner}
\alias{print.multiLearner}
\title{Print an abject of class 'multiLearner'.}
\usage{
\method{print}{multiLearner}(x, ...)
}
\arguments{
\item{x}{Object of class 'multiLearner'}

\item{...}{[\code{any}]\cr}
}
\description{
Print an abject of class 'multiLearner'.
}
