% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multiVarSelect.R
\name{print.multiVarSelect}
\alias{print.multiVarSelect}
\title{Print object of class 'multiVarSelect'.}
\usage{
\method{print}{multiVarSelect}(x, ...)
}
\arguments{
\item{x}{Object of class 'multiVarSelect'.}

\item{...}{[\code{any}]\cr}
}
\description{
Print object of class 'multiVarSelect'.
}
