% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ranger.cved.R
\name{print.ranger.cved}
\alias{print.ranger.cved}
\title{Print object of class 'ranger.cved'.}
\usage{
\method{print}{ranger.cved}(x, ...)
}
\arguments{
\item{x}{Object of class 'ranger.cved'.}

\item{...}{[\code{any}]\cr}
}
\description{
Print object of class 'ranger.cved'.
}
