% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multiLearner.R
\name{print.ranger.layer}
\alias{print.ranger.layer}
\title{Print an abject of class 'ranger.layer'.}
\usage{
\method{print}{ranger.layer}(x, ...)
}
\arguments{
\item{x}{Object of class 'ranger.layer'}

\item{...}{[\code{any}]\cr}
}
\description{
Print an abject of class 'ranger.layer'.
}
