% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.super.lnr.R
\name{print.super.lnr}
\alias{print.super.lnr}
\title{Print object of class 'super.lnr'.}
\usage{
\method{print}{super.lnr}(x, ...)
}
\arguments{
\item{x}{Object of class 'super.lnr'.}

\item{...}{[\code{any}]\cr}
}
\description{
Print object of class 'super.lnr'.
}
