% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.varselectarg.R
\name{print.varselectarg}
\alias{print.varselectarg}
\title{Print object of class varselectarg.}
\usage{
\method{print}{varselectarg}(x, ...)
}
\arguments{
\item{x}{object to be printed.}

\item{...}{More arguments to be passed to further function.}
}
\description{
Print object of class varselectarg.
}
\examples{
test_varselectarg <- varselectarg(object = list(probability = TRUE),
                          layer = "methylation")
test_varselectarg <- add(object = test_varselectarg,
                   layer = "genexpr",
                   list(probability = TRUE))
print(test_varselectarg)
}
