% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.weight.R
\name{print.weight}
\alias{print.weight}
\title{Print weith for weighted final learner.}
\usage{
\method{print}{weight}(x, ...)
}
\arguments{
\item{x}{Object of class 'weoght'.}

\item{...}{[\code{any}]\cr}
}
\description{
Print weith for weighted final learner.
}
