% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger.cved.R
\name{ranger.cved}
\alias{ranger.cved}
\title{Validation and ranger are used to build the final model.}
\usage{
ranger.cved(
  multilrn,
  validation = "createFolds",
  validation_args = list(),
  final_learner = "ranger",
  final_learner_args = list()
)
}
\arguments{
\item{multilrn}{Object containing layer models.}

\item{validation}{Validation function e.g 'createFolds' for cross validation.}

\item{validation_args}{Argumments to be passed to the validation function.}

\item{final_learner}{Name of the final learner.}

\item{final_learner_args}{Arguments to be passer to the final learner.}
}
\value{
An object of class \code{ranger.cved} containing:
\tabular{ll}{
   \code{data} \tab The traing data, e.g. \code{ranger},\cr
   \code{target} \tab The target variable,\cr
   \code{validation} \tab The validation function,\cr
   \code{predict_entities} \tab The predicted entities after validation,\cr
   \code{final_learner} \tab The name of the final learner and \cr
   \code{final_model} \tab The name of the final model.
 }
}
\description{
Validation and ranger are used to build the final model.
}
