% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.methods.R
\name{remove.default}
\alias{remove.default}
\title{Default version of the generic function \code{remove}.}
\usage{
\method{remove}{default}(object, ...)
}
\arguments{
\item{object}{Current object.}

\item{...}{More arguments to be passed to further function.}
}
\description{
Default version of the generic function \code{remove}.
}
