% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.entitty.R
\name{remove.entity}
\alias{remove.entity}
\title{This function removes data for the corresponding layer passed as argument.}
\usage{
\method{remove}{entity}(object, layer, ...)
}
\arguments{
\item{object}{The current entity object.}

\item{layer}{Layer refers to the name of the entity to be removed.}

\item{...}{More arguments to be passed to further function.}
}
\value{
The shorter entity object.
}
\description{
This function removes data for the corresponding layer passed as argument.
}
\examples{
data(entities)
test_entities <- entity(object = entities$methylation,
                          layer = "methylation")
test_entities <- add(object = test_entities,
                   layer = "genexpr",
                   data = entities$genexpr)
test_entities <- remove(object = test_entities,
                      layer = "genexpr")
}
