% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.lrnarg.R,
%   R/miscellaneous.varselectarg.R
\name{remove.lrnarg}
\alias{remove.lrnarg}
\title{This function removes a parameter list for the corresponding layer passed as argument.}
\usage{
\method{remove}{lrnarg}(object, layer, ...)

\method{remove}{lrnarg}(object, layer, ...)
}
\arguments{
\item{object}{The current entity object.}

\item{layer}{Layer refers to the name of the entity to be removed.}

\item{...}{More arguments to be passed to further function.}
}
\value{
The shorter list of arguments.

The shorter list of arguments.
}
\description{
This function removes a parameter list for the corresponding layer passed as argument.

This function removes a parameter list for the corresponding layer passed as argument.
}
\examples{
test_lrnarg <- lrnarg(object = list(probability = TRUE),
                          layer = "methylation")
test_lrnarg <- add(object = test_lrnarg,
                   layer = "genexpr",
                   list(probability = TRUE))
test_lrnarg <- remove(object = test_lrnarg,
                      layer = "genexpr")
test_varselectarg <- varselectarg(object = list(probability = TRUE),
                          layer = "methylation")
test_varselectarg <- add(object = test_varselectarg,
                   layer = "genexpr",
                   list(probability = TRUE))
test_varselectarg <- remove(object = test_varselectarg,
                      layer = "genexpr")
}
