% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.methods.R
\name{remove}
\alias{remove}
\title{Generic function to remove element from the object passed as argument.}
\usage{
remove(object, ...)
}
\arguments{
\item{object}{The given object from class \code{entity}, \code{lnrarg} or \code{varselectarg}.}

\item{...}{More arguments to be passed to further function.}
}
\description{
Generic function to remove element from the object passed as argument.
}
