% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{removeLayer}
\alias{removeLayer}
\title{removeLayer}
\usage{
removeLayer(training, layer_id = NULL)
}
\arguments{
\item{training}{(\code{Training(1)}) \cr
Training object from which the layer will be removed.}

\item{layer_id}{(\code{character(1)}) \cr
ID of the layer to be removed.}
}
\description{
Remove a \link{TrainLayer} or a \link{TrainMetaLayer} from the \link{Training} object passed as argument.
}
