% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{Study}
\alias{Study}
\title{Study Class}
\description{
This class is the basic class of the present package. An object from this class
is designed to contain multiple layers, but only one meta layer.

A study is structured as followed:
\itemize{
\item \link{Layer}: Can be clinical, gene expression, etc.
\itemize{
\item \link{Lrner}: Specific to each layer, it must be set up by the user.
\item \link{TrainData}: Specific to each layer, it must be set up by the user.
\item \link{Model}: Specific to each layer, it is set up by training the learner on the training data.
}
\item \link{MetaLayer}: Basically a \link{Layer}, but with some specific properties.
\itemize{
\item \link{Lrner}: This is the meta learner, it must be set up by the user.
\item \link{TrainData}: Specific to each layer, it is set up internally after cross-validation.
\item \link{Model}: Specific to each layer, it is set up by training the learner on the training data.
}
}

Use the function \code{train} to train a study and \code{predict} to predict
a new study.
}
\seealso{
\link{Layer}
}
\section{Super class}{
\code{\link[fuseMLR:HashTable]{fuseMLR::HashTable}} -> \code{Study}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Study-new}{\code{Study$new()}}
\item \href{#method-Study-print}{\code{Study$print()}}
\item \href{#method-Study-trainLayer}{\code{Study$trainLayer()}}
\item \href{#method-Study-predictLayer}{\code{Study$predictLayer()}}
\item \href{#method-Study-createMetaTrainData}{\code{Study$createMetaTrainData()}}
\item \href{#method-Study-train}{\code{Study$train()}}
\item \href{#method-Study-createMetaNewData}{\code{Study$createMetaNewData()}}
\item \href{#method-Study-predict}{\code{Study$predict()}}
\item \href{#method-Study-getTargetValues}{\code{Study$getTargetValues()}}
\item \href{#method-Study-getIndIDs}{\code{Study$getIndIDs()}}
\item \href{#method-Study-getMetaLayer}{\code{Study$getMetaLayer()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="add2HashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-add2HashTable'><code>fuseMLR::HashTable$add2HashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getFromHashTable'><code>fuseMLR::HashTable$getFromHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getHashTable'><code>fuseMLR::HashTable$getHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getId"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getId'><code>fuseMLR::HashTable$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getKeyClass"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getKeyClass'><code>fuseMLR::HashTable$getKeyClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="removeFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-removeFromHashTable'><code>fuseMLR::HashTable$removeFromHashTable()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-new"></a>}}
\if{latex}{\out{\hypertarget{method-Study-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$new(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
See class Param}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-print"></a>}}
\if{latex}{\out{\hypertarget{method-Study-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-trainLayer"></a>}}
\if{latex}{\out{\hypertarget{method-Study-trainLayer}{}}}
\subsection{Method \code{trainLayer()}}{
Train each layer of the current study.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$trainLayer(ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ind_subset}}{(\code{character()})\cr
Subset of individuals IDs to be used for training.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the object itself, with a model for each layer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-predictLayer"></a>}}
\if{latex}{\out{\hypertarget{method-Study-predictLayer}{}}}
\subsection{Method \code{predictLayer()}}{
Predicts values given new data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$predictLayer(new_study, ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_study}}{(\code{NewData()}) \cr
Object of class \link{NewData}.}

\item{\code{ind_subset}}{(\code{vector()}) \cr
Subset of individuals IDs to be used for training.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \link{Study} with predicted values for each layer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-createMetaTrainData"></a>}}
\if{latex}{\out{\hypertarget{method-Study-createMetaTrainData}{}}}
\subsection{Method \code{createMetaTrainData()}}{
Creates a meta training dataset and assigns it to the meta layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$createMetaTrainData(resampling_method, resampling_arg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resampling_method}}{(\verb{function()}) \cr
Function for internal validation.}

\item{\code{resampling_arg}}{(\code{list()}) \cr
List of arguments to be passed to the function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The current object is returned, with a meta training dataset assigned to the meta layer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-train"></a>}}
\if{latex}{\out{\hypertarget{method-Study-train}{}}}
\subsection{Method \code{train()}}{
Trains the current study. All leaners and the meta learner are trained.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$train(ind_subset = NULL, resampling_method, resampling_arg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ind_subset}}{\code{\link[=vector]{vector()}} \cr
ID subset to be used for training.}

\item{\code{resampling_method}}{(\verb{function()}) \cr
Function for internal validation.}

\item{\code{resampling_arg}}{(\code{list()}) \cr
List of arguments to be passed to the function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The current object is returned, with each learner trained on each layer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-createMetaNewData"></a>}}
\if{latex}{\out{\hypertarget{method-Study-createMetaNewData}{}}}
\subsection{Method \code{createMetaNewData()}}{
Creates a new meta dataset based on layer predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$createMetaNewData(predicted_study)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predicted_study}}{\code{\link[=Study]{Study()}} \cr
Predicted study obtained for each layer by \code{predictLayer()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{NewData} is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Study-predict}{}}}
\subsection{Method \code{predict()}}{
Predicts a new study.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$predict(new_study, ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_study}}{\code{\link[=Study]{Study()}} \cr
A new study to be predicted.}

\item{\code{ind_subset}}{\code{\link[=vector]{vector()}} \cr
Vector of IDs to be predicted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The predicted object. All layers and the meta layer are predicted. This is the final predicted object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-getTargetValues"></a>}}
\if{latex}{\out{\hypertarget{method-Study-getTargetValues}{}}}
\subsection{Method \code{getTargetValues()}}{
Gather target values from all layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$getTargetValues()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing individuals IDs and corresponding target values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-getIndIDs"></a>}}
\if{latex}{\out{\hypertarget{method-Study-getIndIDs}{}}}
\subsection{Method \code{getIndIDs()}}{
Gather individual IDs from all layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$getIndIDs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing individuals IDs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-getMetaLayer"></a>}}
\if{latex}{\out{\hypertarget{method-Study-getMetaLayer}{}}}
\subsection{Method \code{getMetaLayer()}}{
Getter of the meta layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$getMetaLayer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Object from class \link{MetaLayer}
}
}
}
