% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{summary.Training}
\alias{summary.Training}
\title{Training object Summaries}
\usage{
\method{summary}{Training}(object, ...)
}
\arguments{
\item{object}{\code{Training(1)} \cr
The \link{Training} object of interest.}

\item{...}{\cr
Further arguments.}
}
\description{
Summaries a fuseMLR \link{Training} object.
}
