% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_disease.R
\docType{data}
\name{test_disease}
\alias{test_disease}
\title{Simulated test disease for 100 participants approximately balanced into negative and positive groups.}
\format{
A vector of patient status.
}
\usage{
data(test_disease)
}
\description{
A dataset containing simulated entities data using the R package \code{interSIM}.
}
\keyword{datasets}
