% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainStudy.R
\name{TrainStudy}
\alias{TrainStudy}
\title{TrainStudy Class}
\description{
This class is the basic class of the present package. An object from this class
is designed to contain multiple training layers, but only one meta training layer.

The Trainstudy class is structured as followed:
\itemize{
\item \link{TrainLayer}: Specific layer containing:
\itemize{
\item \link{Lrner}: Specific learner. This must be set by the user.
\item \link{TrainData}: Specific training dataset. This must be set up by the user.
\item \link{Model}: Specific model. This is set up by training the learner on the training data.
}
\item \link{TrainMetaLayer}: Basically a \link{TrainLayer}, but with some specific properties.
\itemize{
\item \link{Lrner}: This is the meta learner, it must be set up by the user.
\item \link{TrainData}: Specific meta data. This is set up internally after cross-validation.
\item \link{Model}: Specific meta model. This is set up by training the learner on the training data.
}
}

Use the function \code{train} to train a study and \code{predict} to predict
a new study.
}
\seealso{
\link{TrainLayer}

\link{NewStudy} and \link{PredictStudy}
}
\section{Super class}{
\code{\link[fuseMLR:HashTable]{fuseMLR::HashTable}} -> \code{TrainStudy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TrainStudy-new}{\code{TrainStudy$new()}}
\item \href{#method-TrainStudy-print}{\code{TrainStudy$print()}}
\item \href{#method-TrainStudy-trainLayer}{\code{TrainStudy$trainLayer()}}
\item \href{#method-TrainStudy-predictLayer}{\code{TrainStudy$predictLayer()}}
\item \href{#method-TrainStudy-createMetaTrainData}{\code{TrainStudy$createMetaTrainData()}}
\item \href{#method-TrainStudy-train}{\code{TrainStudy$train()}}
\item \href{#method-TrainStudy-predict}{\code{TrainStudy$predict()}}
\item \href{#method-TrainStudy-getTargetValues}{\code{TrainStudy$getTargetValues()}}
\item \href{#method-TrainStudy-getIndIDs}{\code{TrainStudy$getIndIDs()}}
\item \href{#method-TrainStudy-getTrainMetaLayer}{\code{TrainStudy$getTrainMetaLayer()}}
\item \href{#method-TrainStudy-getIndCol}{\code{TrainStudy$getIndCol()}}
\item \href{#method-TrainStudy-getTarget}{\code{TrainStudy$getTarget()}}
\item \href{#method-TrainStudy-increaseNbTrainedLayer}{\code{TrainStudy$increaseNbTrainedLayer()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="add2HashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-add2HashTable'><code>fuseMLR::HashTable$add2HashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getFromHashTable'><code>fuseMLR::HashTable$getFromHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getHashTable'><code>fuseMLR::HashTable$getHashTable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getId"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getId'><code>fuseMLR::HashTable$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="getKeyClass"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-getKeyClass'><code>fuseMLR::HashTable$getKeyClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fuseMLR" data-topic="HashTable" data-id="removeFromHashTable"><a href='../../fuseMLR/html/HashTable.html#method-HashTable-removeFromHashTable'><code>fuseMLR::HashTable$removeFromHashTable()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-new"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$new(id, ind_col, target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1L)})\cr}

\item{\code{ind_col}}{(\code{character(0L)})
Name of column of individuals IDS}

\item{\code{target}}{(\code{character(0L)})
Name of the target variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-print"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-trainLayer"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-trainLayer}{}}}
\subsection{Method \code{trainLayer()}}{
Train each layer of the current Trainstudy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$trainLayer(ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ind_subset}}{(\code{character()})\cr
Subset of individuals IDs to be used for training.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the object itself, with a model for each layer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-predictLayer"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-predictLayer}{}}}
\subsection{Method \code{predictLayer()}}{
Predicts values given new data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$predictLayer(new_study, ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_study}}{(\code{NewData()}) \cr
Object of class \link{NewData}.}

\item{\code{ind_subset}}{(\code{vector()}) \cr
Subset of individuals IDs to be used for training.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \link{TrainStudy} with predicted values for each layer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-createMetaTrainData"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-createMetaTrainData}{}}}
\subsection{Method \code{createMetaTrainData()}}{
Creates a meta training dataset and assigns it to the meta layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$createMetaTrainData(resampling_method, resampling_arg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resampling_method}}{(\verb{function()}) \cr
Function for internal validation.}

\item{\code{resampling_arg}}{(\code{list()}) \cr
List of arguments to be passed to the function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The current object is returned, with a meta training dataset assigned to the meta layer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-train"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-train}{}}}
\subsection{Method \code{train()}}{
Trains the current study. All leaners and the meta learner are trained.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$train(ind_subset = NULL, resampling_method, resampling_arg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ind_subset}}{\code{\link[=vector]{vector()}} \cr
ID subset to be used for training.}

\item{\code{resampling_method}}{(\verb{function()}) \cr
Function for internal validation.}

\item{\code{resampling_arg}}{(\code{list()}) \cr
List of arguments to be passed to the function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The current object is returned, with each learner trained on each layer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-predict"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-predict}{}}}
\subsection{Method \code{predict()}}{
Predicts a new study.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$predict(new_study, ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_study}}{\link{TrainStudy} \cr
A new study to be predicted.}

\item{\code{ind_subset}}{\code{\link[=vector]{vector()}} \cr
Vector of IDs to be predicted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The predicted object. All layers and the meta layer are predicted. This is the final predicted object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-getTargetValues"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-getTargetValues}{}}}
\subsection{Method \code{getTargetValues()}}{
Gather target values from all layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$getTargetValues()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing individuals IDs and corresponding target values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-getIndIDs"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-getIndIDs}{}}}
\subsection{Method \code{getIndIDs()}}{
Gather individual IDs from all layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$getIndIDs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} containing individuals IDs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-getTrainMetaLayer"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-getTrainMetaLayer}{}}}
\subsection{Method \code{getTrainMetaLayer()}}{
Getter of the meta layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$getTrainMetaLayer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Object from class \link{TrainMetaLayer}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-getIndCol"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-getIndCol}{}}}
\subsection{Method \code{getIndCol()}}{
Getter of the individual column name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$getIndCol()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-getTarget"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-getTarget}{}}}
\subsection{Method \code{getTarget()}}{
Getter of the target variable name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$getTarget()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainStudy-increaseNbTrainedLayer"></a>}}
\if{latex}{\out{\hypertarget{method-TrainStudy-increaseNbTrainedLayer}{}}}
\subsection{Method \code{increaseNbTrainedLayer()}}{
Increase the number of trained layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainStudy$increaseNbTrainedLayer()}\if{html}{\out{</div>}}
}

}
}
