% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTraining.R
\name{upsetplot}
\alias{upsetplot}
\title{An upset plot of overlapping individuals.}
\usage{
upsetplot(object, ...)
}
\arguments{
\item{object}{(\verb{Training(1) or Testing(1)}) \cr
Training or testing object for each the upset plot will be created.}

\item{...}{\cr
Further arguments to be passed to the \code{upset} function from package \code{UpSetR}.}
}
\description{
An upset plot of overlapping individuals.
}
