% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarSel.R
\name{VarSel}
\alias{VarSel}
\title{Varsel Class}
\description{
This class implements a learner. A \link{VarSel} object can only exist as a component of a
\link{TrainLayer} or a \link{TrainMetaLayer} object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VarSel-new}{\code{VarSel$new()}}
\item \href{#method-VarSel-print}{\code{VarSel$print()}}
\item \href{#method-VarSel-summary}{\code{VarSel$summary()}}
\item \href{#method-VarSel-varSelection}{\code{VarSel$varSelection()}}
\item \href{#method-VarSel-getTrainLayer}{\code{VarSel$getTrainLayer()}}
\item \href{#method-VarSel-getId}{\code{VarSel$getId()}}
\item \href{#method-VarSel-getPackage}{\code{VarSel$getPackage()}}
\item \href{#method-VarSel-getVarSubSet}{\code{VarSel$getVarSubSet()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarSel-new"></a>}}
\if{latex}{\out{\hypertarget{method-VarSel-new}{}}}
\subsection{Method \code{new()}}{
Variable selection parameter list.

Learner ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarSel$new(id, package = NULL, varsel_fct, varsel_param, train_layer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)}) \cr
Package that implements the variable selection function.
If NULL, the variable selection  function is called from
the current environment.}

\item{\code{package}}{(\code{character(1)}) \cr
Variable selection function name. Note: Variable selection functions, except \code{Boruta}, must return a vector of selected variables.}

\item{\code{varsel_fct}}{(\code{character(1)}) \cr
Variable selection parameters.}

\item{\code{varsel_param}}{(\code{list(1)}) \cr
Layer on which the learner is stored.}

\item{\code{train_layer}}{(\code{TrainLayer(1)}) \cr
The training layer where to store the learner.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarSel-print"></a>}}
\if{latex}{\out{\hypertarget{method-VarSel-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarSel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarSel-summary"></a>}}
\if{latex}{\out{\hypertarget{method-VarSel-summary}{}}}
\subsection{Method \code{summary()}}{
Summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarSel$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any) \cr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarSel-varSelection"></a>}}
\if{latex}{\out{\hypertarget{method-VarSel-varSelection}{}}}
\subsection{Method \code{varSelection()}}{
Tains the current learner (from class \link{Lrner}) on the current training data (from class \link{TrainData}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarSel$varSelection(ind_subset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ind_subset}}{\code{vector(1)} \cr
Individual ID subset on which the training will be performed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The resulting model, from class \link{Model}, is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarSel-getTrainLayer"></a>}}
\if{latex}{\out{\hypertarget{method-VarSel-getTrainLayer}{}}}
\subsection{Method \code{getTrainLayer()}}{
The current layer is returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarSel$getTrainLayer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link{TrainLayer} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarSel-getId"></a>}}
\if{latex}{\out{\hypertarget{method-VarSel-getId}{}}}
\subsection{Method \code{getId()}}{
Getter of the current learner ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarSel$getId()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current learner ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarSel-getPackage"></a>}}
\if{latex}{\out{\hypertarget{method-VarSel-getPackage}{}}}
\subsection{Method \code{getPackage()}}{
Getter of the variable selection package implementing the variable selection function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarSel$getPackage()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The name of the package implementing the variable selection function.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarSel-getVarSubSet"></a>}}
\if{latex}{\out{\hypertarget{method-VarSel-getVarSubSet}{}}}
\subsection{Method \code{getVarSubSet()}}{
Getter of the list of selected variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarSel$getVarSubSet()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of selected variables..
}
}
}
