% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{varSelection}
\alias{varSelection}
\title{varSelection}
\usage{
varSelection(training, ind_subset = NULL)
}
\arguments{
\item{training}{\code{Training} \cr
Training object for storing the created layer.}

\item{ind_subset}{\code{vector} \cr
ID subset of individuals to be used for variable selection.}
}
\value{
A \code{data.frame} with two columns: layer and selected variables.
}
\description{
Variable selection on the training object passed as argument.
}
\references{
Fouodo, C. J. K., Bleskina, M. & Szymczak, S. fuseMLR: an R package for integrative prediction modeling of multi-omics data. BMC Bioinformatics 26, 221 (2025). https://doi.org/10.1186/s12859-025-06248-4 \cr
}
