% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.mean.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Super learner using weighted mean. This function assigns weights to models.}
\usage{
weighted_mean(multilrn, weight)
}
\arguments{
\item{multilrn}{Model entities.}

\item{weight}{Weights of Model entities.}
}
\value{
A list containing a final learner, a final model and layer models.
}
\description{
Super learner using weighted mean. This function assigns weights to models.
}
