% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingFunctions.R
\name{createTesting}
\alias{createTesting}
\title{createTesting}
\usage{
createTesting(id, ind_col, verbose = TRUE)
}
\arguments{
\item{id}{(\code{character(1)}) \cr
Testing id.}

\item{ind_col}{(\code{character(1)}) \cr
Name of column of individuals IDs in testing data.frame.}

\item{verbose}{(\code{boolean}) \cr
Warning messages will be displayed if set to TRUE.}
}
\value{
A \link{Testing} object.
}
\description{
Creates a \link{Testing} object.
}
