% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTraining.R
\name{createTraining}
\alias{createTraining}
\title{Creates a \link{Training} object.}
\usage{
createTraining(
  id,
  ind_col,
  target,
  target_df,
  problem_type = "classification",
  verbose = TRUE
)
}
\arguments{
\item{id}{(\code{character(1)}) \cr}

\item{ind_col}{(\code{character(1)}) \cr
Name of column of individuals IDS.}

\item{target}{(\code{character(1)}) \cr
Name of the target variable.}

\item{target_df}{(\code{data.frame(1)}) \cr
Data frame with two columns: individual IDs and response variable values.}

\item{problem_type}{(\code{character}) \cr
Either "classification" or "regression".}

\item{verbose}{(\code{boolean}) \cr
Warning messages will be displayed if set to TRUE.}
}
\value{
The created \link{Training} object is returned.
}
\description{
Creates a \link{Training} object.
}
