% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{createTrainLayer}
\alias{createTrainLayer}
\title{createTrainLayer}
\usage{
createTrainLayer(
  training,
  train_layer_id,
  train_data,
  varsel_package = NULL,
  varsel_fct,
  varsel_param = list(),
  lrner_package,
  lrn_fct,
  param_train_list = list(),
  param_pred_list = list(),
  na_rm = TRUE,
  x = "x",
  y = "y",
  object = "object",
  data = "data"
)
}
\arguments{
\item{training}{(\code{Training(1)}) \cr
Training object where the created layer will be stored.}

\item{train_layer_id}{(\code{character(1)}) \cr
ID of the \link{TrainLayer} to be created.}

\item{train_data}{(\code{data.frame(1)}) \cr
Data modality to be stored in \link{TrainData}.}

\item{varsel_package}{(\code{character(1)}) \cr
Name of the package containing the function that implements the variable selection algorithm.\cr}

\item{varsel_fct}{(\code{character(1)}) \cr
Name of the function that performs variable selection.}

\item{varsel_param}{(\code{list}) \cr
List of arguments to be passed to \code{varsel_fct}.}

\item{lrner_package}{(\code{character(1)}) \cr
Name of the package containing the function that implements the learning algorithm.\cr}

\item{lrn_fct}{\verb{(character(1)} \cr
Name of the function that that implements the learning algorithm.}

\item{param_train_list}{\verb{(character(1)} \cr
List of arguments to be passed to \code{lrn_fct}.}

\item{param_pred_list}{\verb{(character(1)} \cr
List of arguments to be passed to \code{predict} when computing predictions.}

\item{na_rm}{\cr
If \code{TRUE}, the individuals with missing predictor values will be removed from the training dataset.}

\item{x}{(\code{character(1)}) \cr
If the name of the argument used by the provided original functions to pass
the matrix of independent variable is not \code{x}, use this argument to specify how it is callled in the provided function.}

\item{y}{(\code{character(1)}) \cr
If the name of the argument used by the provided original functions to pass
the target variable is not \code{y}, use this argument to specify how it is callled in the provided function.}

\item{object}{(\code{character(1)}) \cr
The generic function \code{predict} uses a parameter \code{object} to pass a model.
If the corresponding argument is named differently in your predict function, specify the name.}

\item{data}{(\code{character(1)}) \cr
The generic function \code{predict} uses a parameter \code{data} to pass new data.
If the corresponding argument is named differently in your predict function, specify the name.}
}
\value{
The updated \link{Training} object (with the new layer) is returned.
}
\description{
Creates and store a \link{TrainLayer} on the \link{Training} object passed as argument.
}
