% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTraining.R
\name{createTrainMetaLayer}
\alias{createTrainMetaLayer}
\title{Creates and store a \link{TrainMetaLayer} on the \link{Training} object passed as argument.}
\usage{
createTrainMetaLayer(
  training,
  meta_layer_id,
  lrner_package = NULL,
  lrn_fct,
  param_train_list = list(),
  param_pred_list = list(),
  na_rm = TRUE
)
}
\arguments{
\item{training}{(\code{Training(1)}) \cr
Training object where the created layer will be stored.}

\item{meta_layer_id}{(\code{character(1)}) \cr
ID of the layer to be created.}

\item{lrner_package}{(\code{character(1)}) \cr
Name of the package containing the function that implements the learning algorithm.\cr}

\item{lrn_fct}{\verb{(character(1)} \cr
Name of the function that that implements the learning algorithm.}

\item{param_train_list}{\verb{(character(1)} \cr
List of arguments to be passed to \code{lrn_fct}.}

\item{param_pred_list}{\verb{(character(1)} \cr
List of arguments to be passed to \code{predict} when computing predictions.}

\item{na_rm}{\cr
If \code{TRUE}, the individuals with missing predictor values will be removed from the training dataset.}
}
\value{
The updated \link{Training} object (with the new layer) is returned.
}
\description{
Creates and store a \link{TrainMetaLayer} on the \link{Training} object passed as argument.
}
