% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{fusemlr}
\alias{fusemlr}
\title{fusemlr}
\usage{
fusemlr(
  training,
  ind_subset = NULL,
  use_var_sel = FALSE,
  resampling_method = NULL,
  resampling_arg = list(),
  verbose = TRUE
)
}
\arguments{
\item{training}{(\code{Training(1)}) \cr
Training object where the created layer will be stored.}

\item{ind_subset}{(\code{vector(1)}) \cr
ID subset to be used for training.}

\item{use_var_sel}{\code{boolean(1)} \cr
If TRUE, variable selection is performed before training.}

\item{resampling_method}{(\verb{function(1)}) \cr
Function for internal validation. If not specify, the \code{resampling} function from the package \code{caret} is used for a 10-folds cross-validation.}

\item{resampling_arg}{(\code{list(1)}) \cr
List of arguments to be passed to the function.}

\item{verbose}{(\code{boolean}) \cr
Warning messages will be displayed if set to TRUE.}
}
\value{
The current object is returned, with each learner trained on each layer.
}
\description{
Trains the \link{Training} object passed as argument. All leaners and the meta learner are trained.
}
