% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTraining.R
\name{predict.Training}
\alias{predict.Training}
\title{Computes predictions for the \link{Testing} object passed as argument.}
\usage{
\method{predict}{Training}(object, testing, ind_subset = NULL)
}
\arguments{
\item{object}{(\code{Training(1)}) \cr
Training object to be used to compute predictions.}

\item{testing}{(\code{Testing(1)}) \cr
A new testing object to be predicted.}

\item{ind_subset}{(\code{vector(1)}) \cr
Vector of IDs to be predicted.}
}
\value{
The predicted object. All layers and the meta layer are predicted. This is the final predicted object.
}
\description{
Computes predictions for the \link{Testing} object passed as argument.
}
