% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingFunctions.R
\name{summary.Testing}
\alias{summary.Testing}
\title{Testing object Summaries}
\usage{
\method{summary}{Testing}(object, ...)
}
\arguments{
\item{object}{\code{Testing} \cr
The \link{Testing} object of interest.}

\item{...}{\cr
Further arguments.}
}
\description{
Summaries a \code{fuseMLR} \link{Testing} object.
}
