% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{upsetplot}
\alias{upsetplot}
\title{upsetplot}
\usage{
upsetplot(object, ...)
}
\arguments{
\item{object}{\verb{Training or Testing} \cr
Training or testing object for each the upset plot will be created.}

\item{...}{\code{any} \cr
Further arguments to be passed to the \code{upset} function from package \code{UpSetR}.}
}
\description{
An upset plot of overlapping individuals.
}
