% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{varSelection}
\alias{varSelection}
\title{varSelection}
\usage{
varSelection(training, ind_subset = NULL)
}
\arguments{
\item{training}{\code{Training(1)} \cr
Training object where the created layer will be stored.}

\item{ind_subset}{\code{vector(1)} \cr
ID subset of individuals to be used for variable selection.}
}
\value{
The current layer is returned with the resulting model.
}
\description{
Variable selection on the current training object.
}
