% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingFunctions.R
\name{varSelection}
\alias{varSelection}
\title{varSelection}
\usage{
varSelection(training, ind_subset = NULL)
}
\arguments{
\item{training}{\code{Training} \cr
Training object where the created layer will be stored.}

\item{ind_subset}{\code{vector} \cr
ID subset of individuals to be used for variable selection.}
}
\value{
A \code{data.frame} with two columns: layer and selected variables.
}
\description{
Variable selection on the current training object.
}
\references{
Fouodo C.J.K, Bleskina M. and Szymczak (2024). fuseMLR: An R package for integrative prediction modeling of multi-omics data, paper submitted. \cr
}
