% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedMeanLearner.R
\name{weightedMeanLearner}
\alias{weightedMeanLearner}
\title{The weighted mean super learner}
\usage{
weightedMeanLearner(x, y)
}
\arguments{
\item{x}{\code{data.frame()} \cr
\code{data.frame} of predictors.}

\item{y}{\code{vector} \cr
Target observations.}
}
\value{
A model object of class \code{weightedMeanLeaner}.
}
\description{
Layer weights are estimated for each layer based on Brier score.
}
\examples{
set.seed(20240624)
x = data.frame(x1 = rnorm(50))
y = sample(x = 0:1, size = 50, replace = TRUE)
my_model = weightedMeanLearner(x = x, y = y)

}
