% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedMeanLearner.R
\name{weightedMeanLearner}
\alias{weightedMeanLearner}
\title{The weighted mean meta learner}
\usage{
weightedMeanLearner(x, y, weighted = TRUE)
}
\arguments{
\item{x}{\code{data.frame(1)} \cr
\code{data.frame} of predictors.}

\item{y}{\code{vector(1)} \cr
Target observations. Either binary or two level factor variable.}

\item{weighted}{\cr
If TRUE, the weighted sum is computed.}
}
\value{
A model object of class \code{weightedMeanLeaner}.
}
\description{
Layer weights are estimated for each layer based on Brier score.
}
\examples{
set.seed(20240624L)
x = data.frame(x1 = runif(n = 50L, min = 0, max = 1))
y = sample(x = 0L:1L, size = 50L, replace = TRUE)
my_model = weightedMeanLearner(x = x, y = y)

}
